refactor(api-websockets): migrate GraphQL to CoreGraphQLSchemaFactory pattern
Replace GraphQLSchemaPlugin-based GraphQL setup with a WebsocketsGraphQLFactory
using CoreGraphQLSchemaFactory.createImplementation, matching the convention
used by api-scheduler and other packages. Resolvers now declare DI dependencies
(IdentityContext, use cases) via the builder API instead of manual container resolution.
