+++ packages/api-websockets/src/graphql/feature.ts
@@ -1,0 +1,9 @@
+import { createFeature } from "@webiny/feature/api";
+import { WebsocketsGraphQLFactory } from "./WebsocketsGraphQLFactory.js";
+
+export const WebsocketsGraphQLFactoryFeature = createFeature({
+    name: "WebsocketsGraphQLFactory",
+    register(container) {
+        container.register(WebsocketsGraphQLFactory);
+    }
+});
